requirements.txt
# ==== Core ====
# Streamlit with 3.13対応（または3.12以下でもOK）
streamlit>=1.40.0

# pandas: 3.13は2.2.3以上でwheelあり。3.12以下は2.2.2で安定。
pandas==2.2.2; python_version < '3.13'
pandas>=2.2.3; python_version >= '3.13'

# numpy: 3.13は2.1+を推奨、3.12以下は1.26.4でも可（wheelあり）
numpy>=1.26.4; python_version < '3.13'
numpy>=2.1.0; python_version >= '3.13'

# matplotlib: 3.13は3.9+を推奨。3.12以下は3.8系で安定。
matplotlib==3.8.4; python_version < '3.13'
matplotlib>=3.9.0; python_version >= '3.13'

# pure-Python（wheel不要）
openpyxl>=3.1.2
plotly>=5.24.0
altair>=5.2.0
pydantic>=2.5.0
fpdf2>=2.7.9
python-docx>=0.8.11
